Presidential Escort Gamemode
A Scientist (VIP) spawns with several NTF in the cell. They need escort the VIP to escape from the facility. There are also some SCPs in the facility. If the VIP dies, SCPs win. If the VIP escapes, humans win.

NOTE (because I couldn't find another way to test if the scientist has escaped):
The VIP win condition (aka scientist escapes), is currently based on if the *original* scientist (VIP) is no longer a scientist and possesses a flashlight. The reasoning behind this is 914 could cause faulty win conditions, such as if someone were to turn into a scientist, or if the scientist were to upgrade to an NTF. Because of this possibility, original NTF will spawn without flashlights, so when the scientist escapes and respawns as an upgraded class, they will have a flashlight in their inventory. If the scientist is upgraded and without a flashlight (possible via 914), the round ends with no victory.